move karatsuba module
