+++ src/uint.rs
@@ -20,7 +20,6 @@ mod encoding;
 mod from;
 mod gcd;
 mod inv_mod;
-pub(crate) mod karatsuba_mul;
 pub(crate) mod mul;
 mod mul_mod;
 mod neg;
